Beginnings of work to support Additional Library Manager URLs
See https://github.com/arduino/Arduino/issues/6879

At this point the downloaded additional index files would need to be merged into the main index, that could be done in LibraryInstaller.java and not need to change anything, but mauybe it would be better to do it in LibrariesIndexer.  I don't really know enough Java to continue on this knowing that it probably wouldn't be accepted as a pull request anyway (if I could do this in an hour, the only reason I can figure for it not being done is that there is an actual reticence to have unofficial library repositories).

NB: The preferences were just hacked in, I don't know what design tool or whatever is used to do it properly.